// src/app/core/models/Haghighi/HaghighiPerson.ts
export interface HaghighiPersonModel {
    PersonId?;
    NationalCode?: string;
    FirstName?: string;
    FamilyName?: string;
    BirthDate?: any;
    FatherName?: string;
    //ShenasnamehNumber?: string;
    ShomarehShenasnameh?: string;
    MahalSodoorShenasnamehId?: number;
    SeriShenasnameh?: string;
    SerialShenasnameh?: string;
    // MobileNumber?: string;
    ShomarehMobile?: string;

    CityId?: any;
    CityName?: any;
    ProvinceId?: number;
    ProvinceCode?: any;
    ProvinceName?: any;
}